Delete some classes from corext.refactoring
- remove Checks.java, JDTRefactoringDescriptorComment.java,
  JavaRefactoringArguments.java, JavaRefactoringDescriptorUtil.java,
  SearchResultGroup.java,
